==> data-etl/analyses/05_threats_development.qmd <==
---
title: "Threats: Development"
author: "Jay Matsushiba"
date: "2025/04/25"
format:
  html:
    code-fold: true
---

This vignette describes the first-look exploration and analysis with the BC Assessment data, providing sales history and values of real estate properties throughout British Columbia. The purpose for utilizing this data is to create some spatial layer for modelling development pressures on Bowen Island.

Data citation: BC Assessment, 2020, "BC Assessment Data Advice and Inventory Extracts, 2016-2022", <https://hdl.handle.net/11272.1/AB2/LAPUAB>, Abacus Data Network, V6

```{r setup}
#### Palen Lab Packages ####
devtools::load_all()

#### Other Packages ####
library(here)
library(tidyverse)
library(leaflet)
library(scales)
library(stringr)
library(foreach)
library(sf)
library(units)
library(terra)
library(tidyterra)
library(readxl)
library(googlesheets4)
library(ggplot2)
library(ggnewscale)
library(RColorBrewer)
library(randomcoloR)
library(doParallel)
library(magrittr)

sf_use_s2(FALSE)

# Output Directory
output_dir <- here("output-figures/threats_development")
if(!dir.exists(output_dir)) {
  dir.create(output_dir)
}

# Plots parameters
trails_colour <- "brown"
roads_colour <- "darkgrey"
caption <- paste0("Map created: ", date(), ". Palen Lab. Created by Jay Matsushiba (jmatsush@sfu.ca)")

# Template raster
bowen_mask <- terra::rast(here("data/bowen_mask.tif"))
empty_rast <- terra::rast(
  ext(bowen_mask),
  resolution = res(bowen_mask),
  crs = crs(bowen_mask)
)

# Shoreline 
bowen_shoreline

bowen_mask_sf <- as.polygons(bowen_mask, extent=T) %>%
  st_as_sf() %>%
  st_transform(project_crs) %>%
  st_buffer(1000)

# Create Ocean Polygon
bowen_ocean_sf <- st_difference(st_union(bowen_mask_sf), st_union(bowen_shoreline))
bowen_ocean_linewidth <- 0.5
bowen_ocean_colour <- "#dbdbdc"
```

## BC Assessment Sales Data

This section deals with the data wrangling steps for the Bowen Island sales information. This data wrangling assumes that prices only increase, so lowest prices are associated with the earlier sale dates and vice versa.

```{r bca_sales}
bowen_sales_history <- here("data-raw/bc_assessment/bca_folios_spatial_file_20230103/bca_folios.gpkg.gpkg") %>%
  # Bowen Island Municipality JURISDICTION_CODE is 321
  st_read(query = 'SELECT * FROM WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_SALES_SV WHERE JURISDICTION_CODE LIKE 321') 

# Assumes that first date is lowest price, last date is highest price
bowen_sales_summary  <- bowen_sales_history %>%
  # Drop spatial geometry, join back later
  st_drop_geometry() %>%
  # Remove invalid sales 
  filter(CONVEYANCE_TYPE_DESCRIPTION != "Reject - Not Suitable for Sales Analysis") %>%
  # Group by individual properties, since each transaction separate row
  group_by(FOLIO_ID) %>%
  summarise(first_date = min(CONVEYANCE_DATE),
            last_date = max(CONVEYANCE_DATE),
            lowest_price = min(CONVEYANCE_PRICE),
            highest_price = max(CONVEYANCE_PRICE),
            n_distinct = n_distinct(CONVEYANCE_TYPE_DESCRIPTION) # Show whether land was developed or sold as empty land
            ) %>%
  mutate(change_type = case_when(n_distinct > 1 ~ "changed", .default = "unchanged")) %>% # Show whether land was developed or sold as empty land
  mutate(days_span = as.double(last_date - first_date, units = "days")) %>% # Number of days between first available and most recent sale
  mutate(price_days_slope = (highest_price - lowest_price)/days_span) %>% # Change in price divided by number of days 
  filter(!is.infinite(price_days_slope) & !is.na(price_days_slope)) # Remove invalid price changes (ex. property has no records of being sold in open market)
```

### BC Assessment Line Plot

Each property is one line segment on this plot with start and end points. The start point is the lowest price and first available sale date. The end point is the highest price and most recent sale date.

```{r line_plot}
ggplot(bowen_sales_summary) +
  geom_segment(aes(x = first_date, xend = last_date, y = lowest_price, yend = highest_price, colour = change_type), alpha = 0.2) +
  xlab("Dates") +
  ylab("Price (CAD)")
```

### BC Assessment Interactive Map

Properties that have available change in price over time are plotted on an interactive Leaflet map.

```{r spatial_leaflet}
# Join with spatial data
bowen_folio_addresses <- here("data-raw/bc_assessment/bca_folios_spatial_file_20230103/bca_folios.gpkg.gpkg") %>%
  st_read(query = 'SELECT * FROM WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_ADDRESSES_SV WHERE JURISDICTION_CODE LIKE 321') 

bowen_sales_summary_sf <- merge(bowen_sales_summary, bowen_folio_addresses) %>%
  rename(geometry = geom) %>%
  st_as_sf() %>%
  st_transform(4326)

# LEAFLET
bins <- c(0, 10, 20, 50, 100, 200, 500, 1000, Inf)
pal <- colorBin("YlOrRd", domain = bowen_sales_summary_sf$price_days_slope, bins = bins)
leaflet(bowen_sales_summary_sf) %>%
  addProviderTiles(providers$CartoDB.Positron) %>%
  addPolygons(color = "#444444", weight = 0.5,,
              fillColor = ~pal(price_days_slope),
              fillOpacity = 0.7
              ) %>%
  addLegend(pal = pal, values = ~price_days_slope)
```

## Development Subdivision Potential - Information from Daniel Martin (Bowen Island Municipality Manager of Planning & Development)
This section is on quantifying the development potential on Bowen Island, via subdivisions of land to enable increased development / density. Subdivisions refer to taking larger properties of land and splitting them into smaller properties, which enables increased development and density as it allows more buildings to be built. The current zoning restricts the number of buildings per lot and the minimum area per lot. Therefore, splitting properties that could be made into multiple smaller but above minimum properties increases the number of buildings that can be built. 

Daniel Martin, Manager of Planning and Development at Bowen Island Municipality, provided these spreadsheets with information on the potential for individual lots to be subdivided. Additionally, Martin provided the number of potential further developments possible in Comprehensive Development Zones on Bowen Island. 
```{r}
# RR1 and RR2 (Rural Residential 1 and 2 Zoning)
rr1_rr2 <- here("data-raw/datasets/development_potential_danielmartin/zoning subdivision potential.xlsx") %>%
  readxl::read_xlsx(sheet = "RR1,RR2")
# RR3 (Rural Residential 3 Zoning)
rr3 <- here("data-raw/datasets/development_potential_danielmartin/zoning subdivision potential.xlsx") %>%
  readxl::read_xlsx(sheet = "RR3 properties")
# SR (Settlement Residential zoning)
sr <- here("data-raw/datasets/development_potential_danielmartin/zoning subdivision potential.xlsx") %>%
  readxl::read_xlsx(sheet = "SR properties")
# CD (Comprehensive Development Zones)
## Deal with these separately, need to do some wrangling 

ocp_subdivision <- rbind(rr1_rr2, rr3, sr) %>%
  filter(!is.na(pid))

ocp_subdivision_parcelmap <- merge(ocp_subdivision, bowen_parcelmap, by.x = "pid", by.y = "Name") %>%
  mutate(subdividable = case_when(
    `Can Subdivide?` == 0 ~ FALSE,
    `Can Subdivide?` > 0 ~ TRUE
  )) %>%
  st_as_sf() %>%
  st_transform(project_crs)

ocp_subdivision_parcelmap_f <- ocp_subdivision_parcelmap %>%
  mutate(
    pid = as.numeric(pid)
  )

# Removing duplicated columns
unique(ocp_subdivision_parcelmap_f$pid == ocp_subdivision_parcelmap_f$PID) # TRUE
unique(ocp_subdivision_parcelmap_f$parcelclass == toupper(ocp_subdivision_parcelmap_f$ParcelClass)) # TRUE
unique(ocp_subdivision_parcelmap_f$ownertype == toupper(ocp_subdivision_parcelmap_f$OwnerType)) # FALSE
which(ocp_subdivision_parcelmap_f$ownertype != toupper(ocp_subdivision_parcelmap_f$OwnerType)) # 1763
# turns out its UNCLASS and Unclassified, so the same 
ocp_subdivision_parcelmap_fd <- ocp_subdivision_parcelmap_f %>%
  select(!c(
    ParcelClass, OwnerType, PID
  )) 

# Remove outlier and duplicate PID
unique(ocp_subdivision_parcelmap_fd$pid) %>% length() # 1778, so one fewer than than nrows
ocp_subdivision_parcelmap_fdr <- ocp_subdivision_parcelmap_fd %>%
  filter(`Can Subdivide?` != 46)

write_sf(ocp_subdivision_parcelmap_fdr, here("data-raw/datasets/development_potential_danielmartin/ocp_subdivision_parcelmap.gpkg"), delete_layer = TRUE)

ggplot(ocp_subdivision_parcelmap_fdr) +
  geom_sf(aes(fill = `subdividable`, geometry = geom), color = NA) 
ggplot(ocp_subdivision_parcelmap_fdr) +
  geom_sf(aes(fill = `Can Subdivide?`, geometry = geom), color = NA) +
  scale_fill_viridis_c(direction = -1,
                       breaks = c(0, 10, 20, 40),
                       trans = scales::pseudo_log_trans(sigma = 1))

# Subdividable Map (Binary)
title <- "Bowen Island OCP Subdividable"
subtitle <- "This map shows whether parcels are subdividable into additional units on Bowen Island. Data provided by Daniel Martin (2025)."

subdivide_tf_plot <- bowen_map(raster_layer = empty_rast,
                               title = title,
                               subtitle = subtitle,
                               caption = caption,
                               legend_label = "Subdividable?") +
  geom_sf(data = ocp_subdivision_parcelmap_fdr,
          aes(fill = `subdividable`, geometry = geom), color = NA
  ) +
  geom_sf(data = bowen_ocean_sf,
          linewidth = bowen_ocean_linewidth,
          fill = bowen_ocean_colour
  ) +
  ggplot2::geom_sf(data = bowen_trails,
                   linewidth = 0.3,
                   aes(color = trails_colour)
  ) +
  ggplot2::geom_sf(data = bowen_roads,
                   linewidth = 0.5,
                   aes(color = roads_colour)
  ) +
  scale_colour_identity(name = "",
                        breaks = c(trails_colour, roads_colour),
                        labels = c("Trails", "Roads"),
                        guide = "legend"
  ) +
  ggspatial::annotation_scale(
    location = "br",
    bar_cols = c("grey60", "white")
  ) +
  ggspatial::annotation_north_arrow(
    location = "br", which_north = "true",
    pad_x = unit(0.2, "in"), pad_y = unit(0.4, "in"),
    style = ggspatial::north_arrow_nautical(
      fill = c("grey40", "white"),
      line_col = "grey20"
    )
  ) 

ggplot2::ggsave(
  here(output_dir, "/", paste0(Sys.Date(), "_subdividable_tf.png")),
  subdivide_tf_plot,
  device = ragg::agg_png,
  width = 9, height = 12, units = "in", res = 300
)

# Subdividable Map (Continuous)
title <- "Bowen Island OCP Subdividable Empty Lots"
subtitle <- "This map shows the number of new units that parcels can be subdivided into on Bowen Island. Data provided by Daniel Martin (2025)."

subdivide_cont_plot <- bowen_map(raster_layer = empty_rast,
                               title = title,
                               subtitle = subtitle,
                               caption = caption,
                               legend_label = "Subdividable?") +
  geom_sf(data = ocp_subdivision_parcelmap, aes(fill = `Can Subdivide?`, geometry = geom), color = NA) +
  scale_fill_viridis_c(direction = -1) +
  geom_sf(data = bowen_ocean_sf,
          linewidth = bowen_ocean_linewidth,
          fill = bowen_ocean_colour
  ) +
  ggplot2::geom_sf(data = bowen_trails,
                   linewidth = 0.3,
                   aes(color = trails_colour)
  ) +
  ggplot2::geom_sf(data = bowen_roads,
                   linewidth = 0.5,
                   aes(color = roads_colour)
  ) +
  scale_colour_identity(name = "",
                        breaks = c(trails_colour, roads_colour),
                        labels = c("Trails", "Roads"),
                        guide = "legend"
  ) +
  ggspatial::annotation_scale(
    location = "br",
    bar_cols = c("grey60", "white")
  ) +
  ggspatial::annotation_north_arrow(
    location = "br", which_north = "true",
    pad_x = unit(0.2, "in"), pad_y = unit(0.4, "in"),
    style = ggspatial::north_arrow_nautical(
      fill = c("grey40", "white"),
      line_col = "grey20"
    )
  ) 

ggplot2::ggsave(
  here(output_dir, "/", paste0(Sys.Date(), "_subdividable_cont.png")),
  subdivide_cont_plot,
  device = ragg::agg_png,
  width = 9, height = 12, units = "in", res = 300
)
```
# Biodiversity Values in Subdivideable Properties
```{r}
# TODO: How much relative biodiversity value in subdivideable plots?
# Get Zonation conservation priorities output
# 2025/06/12
# - Ranking by summed conservation value, for all pixels inside each parcel
# - raster fuzzy join between conservation values and parcel, intersect with centroid of raster cell 
# - Only for subdivideable plots 
zonation <- terra::rast(here("inst/extdata/output_zonation/2025-05-05/rankmap.tif"))

# Weighted sum of biodiversity value per parcel
sum_bioval_weighted <- ocp_subdivision_parcelmap_fdr %>%
  select(num_subdivide = `Can Subdivide?`,
         geometry = geom) %>%
  st_cast("MULTIPOLYGON") %>%
  vect() %>%
  terra::extract(zonation, ., 
                 sum, 
                 na.rm = TRUE, 
                 weights = TRUE, 
                 bind = TRUE) # sum biodiversity value by approx fraction of cell covered by polygon

sum_bioval_weighted_sf <- sum_bioval_weighted %>% 
  st_as_sf() %>%
  mutate(area_ha = drop_units(st_area(geometry))/10000,
         potential_units = num_subdivide + 1) %>%
  rename(sum_bioval = rankmap) %>%
  mutate(bioval_by_area = sum_bioval / area_ha,
         potential_units_by_area = potential_units / area_ha) %>%
  mutate(bioval_per_unit = bioval_by_area / potential_units_by_area) %>%
  dplyr::filter(!is.na(sum_bioval)) %>% # Remove tiny parcels 
  st_transform(4326) 
# For web application
st_write(sum_bioval_weighted_sf, here("inst/extdata/6_threats/development_potential.gpkg"), append = F) 


bioval_by_potential_units_gg <- function()  {
  list(geom_sf(data = sum_bioval_weighted_sf, aes(fill = bioval_by_area / potential_units), color = NA),
  scale_fill_viridis_c(name = "Biodiversity / Unit"))
}

test_func <- function() {
  "random"
}

title <- "Bowen Island Relative Biodiversity per Potential Residential Units by Parcel"
subtitle <- "This map shows the relative biodiversity values contained within each vacant parcel, divided by the number of potential new units that these parcels can be subdivided into. Data provided by Daniel Martin (April 2025)."
caption <- paste0("Map created: ", date(), ". Palen Lab. Created by Jay Matsushiba (jmatsush@sfu.ca)")

bioval_by_potential_units_ggplot <- bowen_map_ggplot(bioval_by_potential_units_gg,
                                                     title = title,
                                                     subtitle = subtitle,
                                                     caption = caption
)

ggplot2::ggsave(
  here(output_dir, "/", paste0(Sys.Date(), "_bioval_by_potential_units.png")),
  bioval_by_potential_units_ggplot,
  device = ragg::agg_png,
  width = 9, height = 12, units = "in", res = 300
)


# 2025/06/12
# DONE: create clean outputs, send to Daniel Martin
# DONE: Ask Daniel Martin for vacant lots 
# - Realized that the lots I do have are the vacant lots 
# DONE: Ask Daniel Martin for feedback thinking of subdividable and vacant lots as the most desirable for potential development
# TODO: generally clean up code and document decision making
```


```{r}
parcelmap <- st_read(here("data-raw/datasets/parcelmap/bowen_parcelmap.gpkg")) %>%
  filter(PID != 31243550) # Remove ferry route 
  
parcelmap$geom <- parcelmap$geom %>%
  st_cast("MULTIPOLYGON") %>%
  st_make_valid()

parcelmap$area <- st_area(parcelmap) # equal to SHAPE_Area, units in m^2


biod_vals <- rast(here("output-data/5_values/rankmap.tif"))

parcelmap_vect <- parcelmap %>% 
  vect() %>%
  project(biod_vals)

parcelmap_extract <- terra::extract(biod_vals, parcelmap_vect, sum, bind = T)

parcelmap_sf <- st_as_sf(parcelmap_extract)

ggplot(parcelmap_sf) + geom_sf(aes(fill = rankmap))

```
